Acknowledge message early to avoid repeats
Workaround to not get repeated requests which the server
fails to process.

This means, when the server dies, without answering with any
kind of message, the RPC Request is not re-issued to the server
when it is restarted. In this case, the an RPC client will still
wait for an RPC response. This means in this clase, the client
has to be stopped manually and restarted to resume operation.
This avoids having to remove having to purge RPC request messages
which choke the server.

A probably nicer way to handle this is likely to the store the
in-flight RPC requests in another storage, shared memory or queue
and if the server comes up after having had trouble, it reads
this queue first, and responds to those RPC calls which could
not be processed with an error response (to unblock any client
waiting for a response on them).